Add bindings header for native API functions in nanobind
This commit introduces a new header file `bindings.hpp` within the `ttexalens/cpp/nanobind` directory. The file declares functions that register slices of the native API with the nanobind module. This organization helps maintain readability in the corresponding `bindings.cpp` file by separating the declarations of each binding function.
