+++ ttexalens/cpp/nanobind/bind_dwarf_attribute.cpp
@@ -1,0 +1,438 @@
+// SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
+// SPDX-License-Identifier: Apache-2.0
+
+#include <nanobind/nanobind.h>
+#include <nanobind/stl/string.h>
+#include <nanobind/stl/variant.h>
+#include <nanobind/stl/vector.h>
+
+#include <type_traits>
+#include <variant>
+
+#include "bindings.hpp"
+#include "dwarf_attribute.hpp"
+
+namespace nb = nanobind;
+
+namespace ttexalens::native_elf::bindings {
+
+void bind_dwarf_attribute(nb::module_& m) {
+    // DW_AT_* attribute tags. Use with:
+    //   from ttexalens._native_ttexalens import DwarfAttributeTag as at
+    //   die.get_attribute(at.declaration)
+    nb::enum_<DwarfAttributeTag>(m, "DwarfAttributeTag")
+        .value("abstract_origin", DwarfAttributeTag::abstract_origin)
+        .value("accessibility", DwarfAttributeTag::accessibility)
+        .value("addr_base", DwarfAttributeTag::addr_base)
+        .value("address_class", DwarfAttributeTag::address_class)
+        .value("alignment", DwarfAttributeTag::alignment)
+        .value("allocated", DwarfAttributeTag::allocated)
+        .value("ALTIUM_loclist", DwarfAttributeTag::ALTIUM_loclist)
+        .value("APPLE_block", DwarfAttributeTag::APPLE_block)
+        .value("APPLE_flags", DwarfAttributeTag::APPLE_flags)
+        .value("APPLE_isa", DwarfAttributeTag::APPLE_isa)
+        .value("APPLE_major_runtime_vers", DwarfAttributeTag::APPLE_major_runtime_vers)
+        .value("APPLE_objc_complete_type", DwarfAttributeTag::APPLE_objc_complete_type)
+        .value("APPLE_objc_direct", DwarfAttributeTag::APPLE_objc_direct)
+        .value("APPLE_omit_frame_ptr", DwarfAttributeTag::APPLE_omit_frame_ptr)
+        .value("APPLE_optimized", DwarfAttributeTag::APPLE_optimized)
+        .value("APPLE_origin", DwarfAttributeTag::APPLE_origin)
+        .value("APPLE_property", DwarfAttributeTag::APPLE_property)
+        .value("APPLE_property_attribute", DwarfAttributeTag::APPLE_property_attribute)
+        .value("APPLE_property_getter", DwarfAttributeTag::APPLE_property_getter)
+        .value("APPLE_property_name", DwarfAttributeTag::APPLE_property_name)
+        .value("APPLE_property_setter", DwarfAttributeTag::APPLE_property_setter)
+        .value("APPLE_runtime_class", DwarfAttributeTag::APPLE_runtime_class)
+        .value("APPLE_sdk", DwarfAttributeTag::APPLE_sdk)
+        .value("artificial", DwarfAttributeTag::artificial)
+        .value("associated", DwarfAttributeTag::associated)
+        .value("base_types", DwarfAttributeTag::base_types)
+        .value("binary_scale", DwarfAttributeTag::binary_scale)
+        .value("bit_offset", DwarfAttributeTag::bit_offset)
+        .value("bit_size", DwarfAttributeTag::bit_size)
+        .value("bit_stride", DwarfAttributeTag::bit_stride)
+        .value("body_begin", DwarfAttributeTag::body_begin)
+        .value("body_end", DwarfAttributeTag::body_end)
+        .value("BORLAND_closure", DwarfAttributeTag::BORLAND_closure)
+        .value("BORLAND_Delphi_ABI", DwarfAttributeTag::BORLAND_Delphi_ABI)
+        .value("BORLAND_Delphi_anonymous_method", DwarfAttributeTag::BORLAND_Delphi_anonymous_method)
+        .value("BORLAND_Delphi_class", DwarfAttributeTag::BORLAND_Delphi_class)
+        .value("BORLAND_Delphi_constructor", DwarfAttributeTag::BORLAND_Delphi_constructor)
+        .value("BORLAND_Delphi_destructor", DwarfAttributeTag::BORLAND_Delphi_destructor)
+        .value("BORLAND_Delphi_frameptr", DwarfAttributeTag::BORLAND_Delphi_frameptr)
+        .value("BORLAND_Delphi_interface", DwarfAttributeTag::BORLAND_Delphi_interface)
+        .value("BORLAND_Delphi_metaclass", DwarfAttributeTag::BORLAND_Delphi_metaclass)
+        .value("BORLAND_Delphi_record", DwarfAttributeTag::BORLAND_Delphi_record)
+        .value("BORLAND_Delphi_unit", DwarfAttributeTag::BORLAND_Delphi_unit)
+        .value("BORLAND_property_default", DwarfAttributeTag::BORLAND_property_default)
+        .value("BORLAND_property_implements", DwarfAttributeTag::BORLAND_property_implements)
+        .value("BORLAND_property_index", DwarfAttributeTag::BORLAND_property_index)
+        .value("BORLAND_property_read", DwarfAttributeTag::BORLAND_property_read)
+        .value("BORLAND_property_write", DwarfAttributeTag::BORLAND_property_write)
+        .value("byte_size", DwarfAttributeTag::byte_size)
+        .value("byte_stride", DwarfAttributeTag::byte_stride)
+        .value("call_all_calls", DwarfAttributeTag::call_all_calls)
+        .value("call_all_source_calls", DwarfAttributeTag::call_all_source_calls)
+        .value("call_all_tail_calls", DwarfAttributeTag::call_all_tail_calls)
+        .value("call_column", DwarfAttributeTag::call_column)
+        .value("call_data_location", DwarfAttributeTag::call_data_location)
+        .value("call_data_value", DwarfAttributeTag::call_data_value)
+        .value("call_file", DwarfAttributeTag::call_file)
+        .value("call_line", DwarfAttributeTag::call_line)
+        .value("call_origin", DwarfAttributeTag::call_origin)
+        .value("call_parameter", DwarfAttributeTag::call_parameter)
+        .value("call_pc", DwarfAttributeTag::call_pc)
+        .value("call_return_pc", DwarfAttributeTag::call_return_pc)
+        .value("call_tail_call", DwarfAttributeTag::call_tail_call)
+        .value("call_target", DwarfAttributeTag::call_target)
+        .value("call_target_clobbered", DwarfAttributeTag::call_target_clobbered)
+        .value("call_value", DwarfAttributeTag::call_value)
+        .value("calling_convention", DwarfAttributeTag::calling_convention)
+        .value("common_reference", DwarfAttributeTag::common_reference)
+        .value("comp_dir", DwarfAttributeTag::comp_dir)
+        .value("const_expr", DwarfAttributeTag::const_expr)
+        .value("const_value", DwarfAttributeTag::const_value)
+        .value("containing_type", DwarfAttributeTag::containing_type)
+        .value("count", DwarfAttributeTag::count)
+        .value("CPQ_discontig_ranges", DwarfAttributeTag::CPQ_discontig_ranges)
+        .value("CPQ_prologue_length", DwarfAttributeTag::CPQ_prologue_length)
+        .value("CPQ_semantic_events", DwarfAttributeTag::CPQ_semantic_events)
+        .value("CPQ_split_lifetimes_rtn", DwarfAttributeTag::CPQ_split_lifetimes_rtn)
+        .value("CPQ_split_lifetimes_var", DwarfAttributeTag::CPQ_split_lifetimes_var)
+        .value("data_bit_offset", DwarfAttributeTag::data_bit_offset)
+        .value("data_location", DwarfAttributeTag::data_location)
+        .value("data_member_location", DwarfAttributeTag::data_member_location)
+        .value("decimal_scale", DwarfAttributeTag::decimal_scale)
+        .value("decimal_sign", DwarfAttributeTag::decimal_sign)
+        .value("decl_column", DwarfAttributeTag::decl_column)
+        .value("decl_file", DwarfAttributeTag::decl_file)
+        .value("decl_line", DwarfAttributeTag::decl_line)
+        .value("declaration", DwarfAttributeTag::declaration)
+        .value("default_value", DwarfAttributeTag::default_value)
+        .value("defaulted", DwarfAttributeTag::defaulted)
+        .value("deleted", DwarfAttributeTag::deleted)
+        .value("description", DwarfAttributeTag::description)
+        .value("digit_count", DwarfAttributeTag::digit_count)
+        .value("discr", DwarfAttributeTag::discr)
+        .value("discr_list", DwarfAttributeTag::discr_list)
+        .value("discr_value", DwarfAttributeTag::discr_value)
+        .value("dwo_id", DwarfAttributeTag::dwo_id)
+        .value("dwo_name", DwarfAttributeTag::dwo_name)
+        .value("element_list", DwarfAttributeTag::element_list)
+        .value("elemental", DwarfAttributeTag::elemental)
+        .value("encoding", DwarfAttributeTag::encoding)
+        .value("endianity", DwarfAttributeTag::endianity)
+        .value("entry_pc", DwarfAttributeTag::entry_pc)
+        .value("enum_class", DwarfAttributeTag::enum_class)
+        .value("explicit", DwarfAttributeTag::explicit_)
+        .value("export_symbols", DwarfAttributeTag::export_symbols)
+        .value("extension", DwarfAttributeTag::extension)
+        .value("external", DwarfAttributeTag::external)
+        .value("frame_base", DwarfAttributeTag::frame_base)
+        .value("friend", DwarfAttributeTag::friend_)
+        .value("ghs_frames", DwarfAttributeTag::ghs_frames)
+        .value("ghs_frsm", DwarfAttributeTag::ghs_frsm)
+        .value("ghs_lbrace_line", DwarfAttributeTag::ghs_lbrace_line)
+        .value("ghs_mangled", DwarfAttributeTag::ghs_mangled)
+        .value("ghs_namespace_alias", DwarfAttributeTag::ghs_namespace_alias)
+        .value("ghs_rsm", DwarfAttributeTag::ghs_rsm)
+        .value("ghs_rso", DwarfAttributeTag::ghs_rso)
+        .value("ghs_subcpu", DwarfAttributeTag::ghs_subcpu)
+        .value("ghs_using_declaration", DwarfAttributeTag::ghs_using_declaration)
+        .value("ghs_using_namespace", DwarfAttributeTag::ghs_using_namespace)
+        .value("GNAT_descriptive_type", DwarfAttributeTag::GNAT_descriptive_type)
+        .value("GNU_addr_base", DwarfAttributeTag::GNU_addr_base)
+        .value("GNU_all_call_sites", DwarfAttributeTag::GNU_all_call_sites)
+        .value("GNU_all_source_call_sites", DwarfAttributeTag::GNU_all_source_call_sites)
+        .value("GNU_all_tail_call_sites", DwarfAttributeTag::GNU_all_tail_call_sites)
+        .value("GNU_bias", DwarfAttributeTag::GNU_bias)
+        .value("GNU_call_site_data_value", DwarfAttributeTag::GNU_call_site_data_value)
+        .value("GNU_call_site_target", DwarfAttributeTag::GNU_call_site_target)
+        .value("GNU_call_site_target_clobbered", DwarfAttributeTag::GNU_call_site_target_clobbered)
+        .value("GNU_call_site_value", DwarfAttributeTag::GNU_call_site_value)
+        .value("GNU_deleted", DwarfAttributeTag::GNU_deleted)
+        .value("GNU_denominator", DwarfAttributeTag::GNU_denominator)
+        .value("GNU_discriminator", DwarfAttributeTag::GNU_discriminator)
+        .value("GNU_dwo_id", DwarfAttributeTag::GNU_dwo_id)
+        .value("GNU_dwo_name", DwarfAttributeTag::GNU_dwo_name)
+        .value("GNU_entry_view", DwarfAttributeTag::GNU_entry_view)
+        .value("GNU_exclusive_locks_required", DwarfAttributeTag::GNU_exclusive_locks_required)
+        .value("GNU_guarded", DwarfAttributeTag::GNU_guarded)
+        .value("GNU_guarded_by", DwarfAttributeTag::GNU_guarded_by)
+        .value("GNU_locks_excluded", DwarfAttributeTag::GNU_locks_excluded)
+        .value("GNU_locviews", DwarfAttributeTag::GNU_locviews)
+        .value("GNU_macros", DwarfAttributeTag::GNU_macros)
+        .value("GNU_numerator", DwarfAttributeTag::GNU_numerator)
+        .value("GNU_odr_signature", DwarfAttributeTag::GNU_odr_signature)
+        .value("GNU_pt_guarded", DwarfAttributeTag::GNU_pt_guarded)
+        .value("GNU_pt_guarded_by", DwarfAttributeTag::GNU_pt_guarded_by)
+        .value("GNU_pubnames", DwarfAttributeTag::GNU_pubnames)
+        .value("GNU_pubtypes", DwarfAttributeTag::GNU_pubtypes)
+        .value("GNU_ranges_base", DwarfAttributeTag::GNU_ranges_base)
+        .value("GNU_shared_locks_required", DwarfAttributeTag::GNU_shared_locks_required)
+        .value("GNU_tail_call", DwarfAttributeTag::GNU_tail_call)
+        .value("GNU_template_name", DwarfAttributeTag::GNU_template_name)
+        .value("GNU_vector", DwarfAttributeTag::GNU_vector)
+        .value("go_closure_offset", DwarfAttributeTag::go_closure_offset)
+        .value("go_dict_index", DwarfAttributeTag::go_dict_index)
+        .value("go_elem", DwarfAttributeTag::go_elem)
+        .value("go_embedded_field", DwarfAttributeTag::go_embedded_field)
+        .value("go_key", DwarfAttributeTag::go_key)
+        .value("go_kind", DwarfAttributeTag::go_kind)
+        .value("go_package_name", DwarfAttributeTag::go_package_name)
+        .value("go_runtime_type", DwarfAttributeTag::go_runtime_type)
+        .value("high_pc", DwarfAttributeTag::high_pc)
+        .value("HP_actuals_stmt_list", DwarfAttributeTag::HP_actuals_stmt_list)
+        .value("HP_all_variables_modifiable", DwarfAttributeTag::HP_all_variables_modifiable)
+        .value("HP_block_index", DwarfAttributeTag::HP_block_index)
+        .value("HP_cold_region_high_pc", DwarfAttributeTag::HP_cold_region_high_pc)
+        .value("HP_cold_region_low_pc", DwarfAttributeTag::HP_cold_region_low_pc)
+        .value("HP_default_location", DwarfAttributeTag::HP_default_location)
+        .value("HP_definition_points", DwarfAttributeTag::HP_definition_points)
+        .value("HP_epilogue", DwarfAttributeTag::HP_epilogue)
+        .value("HP_is_result_param", DwarfAttributeTag::HP_is_result_param)
+        .value("HP_linkage_name", DwarfAttributeTag::HP_linkage_name)
+        .value("HP_opt_flags", DwarfAttributeTag::HP_opt_flags)
+        .value("HP_opt_level", DwarfAttributeTag::HP_opt_level)
+        .value("HP_pass_by_reference", DwarfAttributeTag::HP_pass_by_reference)
+        .value("HP_proc_per_section", DwarfAttributeTag::HP_proc_per_section)
+        .value("HP_prof_flags", DwarfAttributeTag::HP_prof_flags)
+        .value("HP_prof_version_id", DwarfAttributeTag::HP_prof_version_id)
+        .value("HP_prologue", DwarfAttributeTag::HP_prologue)
+        .value("HP_raw_data_ptr", DwarfAttributeTag::HP_raw_data_ptr)
+        .value("HP_unit_name", DwarfAttributeTag::HP_unit_name)
+        .value("HP_unit_size", DwarfAttributeTag::HP_unit_size)
+        .value("HP_unmodifiable", DwarfAttributeTag::HP_unmodifiable)
+        .value("HP_widened_byte_size", DwarfAttributeTag::HP_widened_byte_size)
+        .value("IBM_alt_srcview", DwarfAttributeTag::IBM_alt_srcview)
+        .value("IBM_home_location", DwarfAttributeTag::IBM_home_location)
+        .value("IBM_wsa_addr", DwarfAttributeTag::IBM_wsa_addr)
+        .value("identifier_case", DwarfAttributeTag::identifier_case)
+        .value("import_", DwarfAttributeTag::import)
+        .value("inline", DwarfAttributeTag::inline_)
+        .value("INTEL_other_endian", DwarfAttributeTag::INTEL_other_endian)
+        .value("is_optional", DwarfAttributeTag::is_optional)
+        .value("language", DwarfAttributeTag::language)
+        .value("language_name", DwarfAttributeTag::language_name)
+        .value("language_version", DwarfAttributeTag::language_version)
+        .value("linkage_name", DwarfAttributeTag::linkage_name)
+        .value("LLVM_active_lane", DwarfAttributeTag::LLVM_active_lane)
+        .value("LLVM_apinotes", DwarfAttributeTag::LLVM_apinotes)
+        .value("LLVM_augmentation", DwarfAttributeTag::LLVM_augmentation)
+        .value("LLVM_config_macros", DwarfAttributeTag::LLVM_config_macros)
+        .value("LLVM_include_path", DwarfAttributeTag::LLVM_include_path)
+        .value("LLVM_lane_pc", DwarfAttributeTag::LLVM_lane_pc)
+        .value("LLVM_lanes", DwarfAttributeTag::LLVM_lanes)
+        .value("LLVM_sysroot", DwarfAttributeTag::LLVM_sysroot)
+        .value("LLVM_tag_offset", DwarfAttributeTag::LLVM_tag_offset)
+        .value("LLVM_vector_size", DwarfAttributeTag::LLVM_vector_size)
+        .value("location", DwarfAttributeTag::location)
+        .value("loclists_base", DwarfAttributeTag::loclists_base)
+        .value("low_pc", DwarfAttributeTag::low_pc)
+        .value("lower_bound", DwarfAttributeTag::lower_bound)
+        .value("mac_info", DwarfAttributeTag::mac_info)
+        .value("macro_info", DwarfAttributeTag::macro_info)
+        .value("macros", DwarfAttributeTag::macros)
+        .value("main_subprogram", DwarfAttributeTag::main_subprogram)
+        .value("member", DwarfAttributeTag::member)
+        .value("MIPS_abstract_name", DwarfAttributeTag::MIPS_abstract_name)
+        .value("MIPS_allocatable_dopetype", DwarfAttributeTag::MIPS_allocatable_dopetype)
+        .value("MIPS_assumed_shape_dopetype", DwarfAttributeTag::MIPS_assumed_shape_dopetype)
+        .value("MIPS_assumed_size", DwarfAttributeTag::MIPS_assumed_size)
+        .value("MIPS_clone_origin", DwarfAttributeTag::MIPS_clone_origin)
+        .value("MIPS_epilog_begin", DwarfAttributeTag::MIPS_epilog_begin)
+        .value("MIPS_fde", DwarfAttributeTag::MIPS_fde)
+        .value("MIPS_has_inlines", DwarfAttributeTag::MIPS_has_inlines)
+        .value("MIPS_linkage_name", DwarfAttributeTag::MIPS_linkage_name)
+        .value("MIPS_loop_begin", DwarfAttributeTag::MIPS_loop_begin)
+        .value("MIPS_loop_unroll_factor", DwarfAttributeTag::MIPS_loop_unroll_factor)
+        .value("MIPS_ptr_dopetype", DwarfAttributeTag::MIPS_ptr_dopetype)
+        .value("MIPS_software_pipeline_depth", DwarfAttributeTag::MIPS_software_pipeline_depth)
+        .value("MIPS_stride", DwarfAttributeTag::MIPS_stride)
+        .value("MIPS_stride_byte", DwarfAttributeTag::MIPS_stride_byte)
+        .value("MIPS_stride_elem", DwarfAttributeTag::MIPS_stride_elem)
+        .value("MIPS_tail_loop_begin", DwarfAttributeTag::MIPS_tail_loop_begin)
+        .value("mutable", DwarfAttributeTag::mutable_)
+        .value("name_", DwarfAttributeTag::name)
+        .value("namelist_item", DwarfAttributeTag::namelist_item)
+        .value("noreturn", DwarfAttributeTag::noreturn)
+        .value("object_pointer", DwarfAttributeTag::object_pointer)
+        .value("ordering", DwarfAttributeTag::ordering)
+        .value("PGI_lbase", DwarfAttributeTag::PGI_lbase)
+        .value("PGI_lstride", DwarfAttributeTag::PGI_lstride)
+        .value("PGI_soffset", DwarfAttributeTag::PGI_soffset)
+        .value("picture_string", DwarfAttributeTag::picture_string)
+        .value("priority", DwarfAttributeTag::priority)
+        .value("producer", DwarfAttributeTag::producer)
+        .value("prototyped", DwarfAttributeTag::prototyped)
+        .value("pure", DwarfAttributeTag::pure)
+        .value("ranges", DwarfAttributeTag::ranges)
+        .value("rank", DwarfAttributeTag::rank)
+        .value("recursive", DwarfAttributeTag::recursive)
+        .value("reference", DwarfAttributeTag::reference)
+        .value("return_addr", DwarfAttributeTag::return_addr)
+        .value("rnglists_base", DwarfAttributeTag::rnglists_base)
+        .value("rvalue_reference", DwarfAttributeTag::rvalue_reference)
+        .value("segment", DwarfAttributeTag::segment)
+        .value("sf_names", DwarfAttributeTag::sf_names)
+        .value("sibling", DwarfAttributeTag::sibling)
+        .value("signature", DwarfAttributeTag::signature)
+        .value("small", DwarfAttributeTag::small)
+        .value("specification", DwarfAttributeTag::specification)
+        .value("src_coords", DwarfAttributeTag::src_coords)
+        .value("src_info", DwarfAttributeTag::src_info)
+        .value("start_scope", DwarfAttributeTag::start_scope)
+        .value("static_link", DwarfAttributeTag::static_link)
+        .value("stmt_list", DwarfAttributeTag::stmt_list)
+        .value("str_offsets_base", DwarfAttributeTag::str_offsets_base)
+        .value("stride", DwarfAttributeTag::stride)
+        .value("stride_size", DwarfAttributeTag::stride_size)
+        .value("string_length", DwarfAttributeTag::string_length)
+        .value("string_length_bit_size", DwarfAttributeTag::string_length_bit_size)
+        .value("string_length_byte_size", DwarfAttributeTag::string_length_byte_size)
+        .value("SUN_alignment", DwarfAttributeTag::SUN_alignment)
+        .value("SUN_amd64_parmdump", DwarfAttributeTag::SUN_amd64_parmdump)
+        .value("SUN_browser_file", DwarfAttributeTag::SUN_browser_file)
+        .value("SUN_c_vla", DwarfAttributeTag::SUN_c_vla)
+        .value("SUN_cf_kind", DwarfAttributeTag::SUN_cf_kind)
+        .value("SUN_command_line", DwarfAttributeTag::SUN_command_line)
+        .value("SUN_compile_options", DwarfAttributeTag::SUN_compile_options)
+        .value("SUN_count_guarantee", DwarfAttributeTag::SUN_count_guarantee)
+        .value("SUN_dtor_length", DwarfAttributeTag::SUN_dtor_length)
+        .value("SUN_dtor_start", DwarfAttributeTag::SUN_dtor_start)
+        .value("SUN_dtor_state_deltas", DwarfAttributeTag::SUN_dtor_state_deltas)
+        .value("SUN_dtor_state_final", DwarfAttributeTag::SUN_dtor_state_final)
+        .value("SUN_dtor_state_initial", DwarfAttributeTag::SUN_dtor_state_initial)
+        .value("SUN_f90_allocatable", DwarfAttributeTag::SUN_f90_allocatable)
+        .value("SUN_f90_assumed_shape_array", DwarfAttributeTag::SUN_f90_assumed_shape_array)
+        .value("SUN_f90_pointer", DwarfAttributeTag::SUN_f90_pointer)
+        .value("SUN_f90_use_only", DwarfAttributeTag::SUN_f90_use_only)
+        .value("SUN_fortran_based", DwarfAttributeTag::SUN_fortran_based)
+        .value("SUN_fortran_main_alias", DwarfAttributeTag::SUN_fortran_main_alias)
+        .value("SUN_func_offset", DwarfAttributeTag::SUN_func_offset)
+        .value("SUN_func_offsets", DwarfAttributeTag::SUN_func_offsets)
+        .value("SUN_hwcprof_signature", DwarfAttributeTag::SUN_hwcprof_signature)
+        .value("SUN_import_by_lname", DwarfAttributeTag::SUN_import_by_lname)
+        .value("SUN_import_by_name", DwarfAttributeTag::SUN_import_by_name)
+        .value("SUN_is_omp_child_func", DwarfAttributeTag::SUN_is_omp_child_func)
+        .value("SUN_language", DwarfAttributeTag::SUN_language)
+        .value("SUN_link_name", DwarfAttributeTag::SUN_link_name)
+        .value("SUN_memop_signature", DwarfAttributeTag::SUN_memop_signature)
+        .value("SUN_memop_type_ref", DwarfAttributeTag::SUN_memop_type_ref)
+        .value("SUN_namelist_spec", DwarfAttributeTag::SUN_namelist_spec)
+        .value("SUN_obj_dir", DwarfAttributeTag::SUN_obj_dir)
+        .value("SUN_obj_file", DwarfAttributeTag::SUN_obj_file)
+        .value("SUN_omp_child_func", DwarfAttributeTag::SUN_omp_child_func)
+        .value("SUN_omp_tpriv_addr", DwarfAttributeTag::SUN_omp_tpriv_addr)
+        .value("SUN_original_name", DwarfAttributeTag::SUN_original_name)
+        .value("SUN_part_link_name", DwarfAttributeTag::SUN_part_link_name)
+        .value("SUN_pass_by_ref", DwarfAttributeTag::SUN_pass_by_ref)
+        .value("SUN_pass_with_const", DwarfAttributeTag::SUN_pass_with_const)
+        .value("SUN_profile_id", DwarfAttributeTag::SUN_profile_id)
+        .value("SUN_return_value_ptr", DwarfAttributeTag::SUN_return_value_ptr)
+        .value("SUN_return_with_const", DwarfAttributeTag::SUN_return_with_const)
+        .value("SUN_template", DwarfAttributeTag::SUN_template)
+        .value("SUN_vbase", DwarfAttributeTag::SUN_vbase)
+        .value("SUN_vtable", DwarfAttributeTag::SUN_vtable)
+        .value("SUN_vtable_abi", DwarfAttributeTag::SUN_vtable_abi)
+        .value("SUN_vtable_index", DwarfAttributeTag::SUN_vtable_index)
+        .value("threads_scaled", DwarfAttributeTag::threads_scaled)
+        .value("TI_asm", DwarfAttributeTag::TI_asm)
+        .value("TI_interrupt", DwarfAttributeTag::TI_interrupt)
+        .value("TI_skeletal", DwarfAttributeTag::TI_skeletal)
+        .value("TI_symbol_name", DwarfAttributeTag::TI_symbol_name)
+        .value("TI_veneer", DwarfAttributeTag::TI_veneer)
+        .value("TI_version", DwarfAttributeTag::TI_version)
+        .value("trampoline", DwarfAttributeTag::trampoline)
+        .value("type", DwarfAttributeTag::type)
+        .value("upc_threads_scaled", DwarfAttributeTag::upc_threads_scaled)
+        .value("upper_bound", DwarfAttributeTag::upper_bound)
+        .value("use_GNAT_descriptive_type", DwarfAttributeTag::use_GNAT_descriptive_type)
+        .value("use_location", DwarfAttributeTag::use_location)
+        .value("use_UTF8", DwarfAttributeTag::use_UTF8)
+        .value("variable_parameter", DwarfAttributeTag::variable_parameter)
+        .value("virtuality", DwarfAttributeTag::virtuality)
+        .value("visibility", DwarfAttributeTag::visibility)
+        .value("VMS_rtnbeg_pd_address", DwarfAttributeTag::VMS_rtnbeg_pd_address)
+        .value("vtable_elem_location", DwarfAttributeTag::vtable_elem_location);
+
+    // DW_FORM_* — attribute form. Determines which std::variant alternative
+    // DwarfAttribute.value will hold (see dwarf_attribute.hpp).
+    nb::enum_<DwarfAttributeForm>(m, "DwarfAttributeForm")
+        .value("addr", DwarfAttributeForm::addr)
+        .value("addrx", DwarfAttributeForm::addrx)
+        .value("addrx1", DwarfAttributeForm::addrx1)
+        .value("addrx2", DwarfAttributeForm::addrx2)
+        .value("addrx3", DwarfAttributeForm::addrx3)
+        .value("addrx4", DwarfAttributeForm::addrx4)
+        .value("block", DwarfAttributeForm::block)
+        .value("block1", DwarfAttributeForm::block1)
+        .value("block2", DwarfAttributeForm::block2)
+        .value("block4", DwarfAttributeForm::block4)
+        .value("data1", DwarfAttributeForm::data1)
+        .value("data16", DwarfAttributeForm::data16)
+        .value("data2", DwarfAttributeForm::data2)
+        .value("data4", DwarfAttributeForm::data4)
+        .value("data8", DwarfAttributeForm::data8)
+        .value("exprloc", DwarfAttributeForm::exprloc)
+        .value("flag", DwarfAttributeForm::flag)
+        .value("flag_present", DwarfAttributeForm::flag_present)
+        .value("GNU_addr_index", DwarfAttributeForm::GNU_addr_index)
+        .value("GNU_ref_alt", DwarfAttributeForm::GNU_ref_alt)
+        .value("GNU_str_index", DwarfAttributeForm::GNU_str_index)
+        .value("GNU_strp_alt", DwarfAttributeForm::GNU_strp_alt)
+        .value("implicit_const", DwarfAttributeForm::implicit_const)
+        .value("indirect", DwarfAttributeForm::indirect)
+        .value("line_strp", DwarfAttributeForm::line_strp)
+        .value("LLVM_addrx_offset", DwarfAttributeForm::LLVM_addrx_offset)
+        .value("loclistx", DwarfAttributeForm::loclistx)
+        .value("ref1", DwarfAttributeForm::ref1)
+        .value("ref2", DwarfAttributeForm::ref2)
+        .value("ref4", DwarfAttributeForm::ref4)
+        .value("ref8", DwarfAttributeForm::ref8)
+        .value("ref_addr", DwarfAttributeForm::ref_addr)
+        .value("ref_sig8", DwarfAttributeForm::ref_sig8)
+        .value("ref_sup4", DwarfAttributeForm::ref_sup4)
+        .value("ref_sup8", DwarfAttributeForm::ref_sup8)
+        .value("ref_udata", DwarfAttributeForm::ref_udata)
+        .value("rnglistx", DwarfAttributeForm::rnglistx)
+        .value("sdata", DwarfAttributeForm::sdata)
+        .value("sec_offset", DwarfAttributeForm::sec_offset)
+        .value("string", DwarfAttributeForm::string)
+        .value("strp", DwarfAttributeForm::strp)
+        .value("strp_sup", DwarfAttributeForm::strp_sup)
+        .value("strx", DwarfAttributeForm::strx)
+        .value("strx1", DwarfAttributeForm::strx1)
+        .value("strx2", DwarfAttributeForm::strx2)
+        .value("strx3", DwarfAttributeForm::strx3)
+        .value("strx4", DwarfAttributeForm::strx4)
+        .value("udata", DwarfAttributeForm::udata);
+
+    nb::class_<DwarfAttribute>(m, "DwarfAttribute")
+        .def_prop_ro("tag", &DwarfAttribute::get_tag)
+        .def_prop_ro("form", &DwarfAttribute::get_form)
+        // Variant alternatives map to bool / int / str / bytes / None. Default
+        // nanobind would expose std::vector<uint8_t> as a Python list[int],
+        // but DW_FORM_block* / exprloc payloads are conceptually opaque byte
+        // strings — hand them out as nb::bytes so callers don't have to
+        // wrap with bytes(...) before struct.unpack / hash / etc.
+        .def_prop_ro(
+            "value",
+            [](const DwarfAttribute& a) -> nb::object {
+                return std::visit(
+                    [](const auto& v) -> nb::object {
+                        using T = std::decay_t<decltype(v)>;
+                        if constexpr (std::is_same_v<T, std::monostate>) {
+                            return nb::none();
+                        } else if constexpr (std::is_same_v<T, std::vector<uint8_t>>) {
+                            return nb::bytes(reinterpret_cast<const char*>(v.data()), v.size());
+                        } else {
+                            return nb::cast(v);  // bool, ints, str
+                        }
+                    },
+                    a.get_value());
+            },
+            nb::sig("def value(self) -> bool | int | str | bytes | None"));
+}
+
+}  // namespace ttexalens::native_elf::bindings
